End-to-End-Fraud-Detection-Project/requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
streamlit
imbalanced-learn
lazypredict
pillow
